+++ cmake/RobotologySuperbuildLogic.cmake
@@ -77,7 +77,7 @@ endif()
 if(ROBOTOLOGY_ENABLE_HUMAN_DYNAMICS)
   find_or_build_package(forcetorque-yarp-devices)
   find_or_build_package(wearables)
-  find_or_build_package(human-dynamics-estimation)
+  find_or_build_package(HumanDynamicsEstimation)
   find_or_build_package(human-gazebo)
 endif()